Refactor test utility package to `container`
Renamed `utils` package used in integration tests to `container` for better clarity and organization. Updated all relevant imports and function calls across affected files to reflect the change.
