+++ tests/integration/container/test_container.go
@@ -1,4 +1,4 @@
-package utils
+package container
 
 import (
 	"context"
